+++ client/src/views/Sshkey.vue
@@ -2,7 +2,7 @@
   <section v-if="isAdmin" class="section">
     <div class="container">
       <h1 class="title has-text-info"><font-awesome-icon icon="key" /> SSH Key</h1>
-      <BulmaTextArea v-if="update" v-model="ssh.key" label="Private Key" placeholder="-----BEGIN RSA PRIVATE KEY-----" :hasError="$v.ssh.key.$invalid" :errors="[]" />
+      <BulmaTextArea v-if="update" v-model="ssh.key" label="Private Key" placeholder="-----BEGIN RSA PRIVATE KEY-----" :hasError="$v.ssh.key.$invalid" :errors="[{if:$v.ssh.key.$invalid,label:'Enter a valid private key followed by a new line'}]" />
       <div class="field"  v-if="!update">
         <label class="label">Private Key</label>
         <p @click="update=true"  class="box is-clickable is-family-monospace enable-line-break is-size-7">{{ ssh.art }}</p>
@@ -26,7 +26,7 @@
   import BulmaTextArea from './../components/BulmaTextArea.vue'
   import TokenStorage from './../lib/TokenStorage'
   import { required, email, minValue,maxValue,minLength,maxLength,helpers,requiredIf,sameAs } from 'vuelidate/lib/validators'
-  const privatekey = helpers.regex("privatekey",/^-----BEGIN RSA PRIVATE KEY-----([^-!]+)-----END RSA PRIVATE KEY-----\s*$/gm)
+  const privatekey = helpers.regex("privatekey",/^-----BEGIN (.+) PRIVATE KEY-----([^-!]+)-----END \1 PRIVATE KEY-----\r?\n$/gm)
   Vue.use(Vuelidate)
 
   export default{