+++ server/src/models/repo.model.js
@@ -65,7 +65,7 @@ Repo.findByName = function (name,text,callback) {
       callback(e)
       return false
     }
-    var clone = exec(`git remote show origin && git log -n 1`,{cwd:directory})
+    var info = exec(`git remote show origin && git log -n 1`,{cwd:directory})
     var output=[]
     if(!text){
       output.push(`<h1 class='subtitle'>${directory}</h1>`)
@@ -73,12 +73,12 @@ Repo.findByName = function (name,text,callback) {
       output.push(directory)
     }
 
-    clone.stdout.on('data', function(a){
+    info.stdout.on('data', function(a){
       // logger.sill(a)
       output.push(a)
     });
 
-    clone.on('exit',function(code){
+    info.on('exit',function(code){
       logger.debug('exit')
       if(code==0){
         if(!text){
@@ -91,7 +91,7 @@ Repo.findByName = function (name,text,callback) {
       }
     });
 
-    clone.stderr.on('data',function(a){
+    info.stderr.on('data',function(a){
       output.push(a)
       logger.error('stderr:'+a);
     });
@@ -102,9 +102,9 @@ Repo.findByName = function (name,text,callback) {
 }
 
 // run git clone
-Repo.create = function (repo, callback) {
+Repo.create = function (uri,command, callback) {
   try{
-    logger.notice("Creating repository " + repo)
+    logger.notice("Creating repository " + (uri)?uri:command)
     var directory = config.repoPath
     try{
       fs.accessSync(directory)
@@ -118,7 +118,17 @@ Repo.create = function (repo, callback) {
         return false
       }
     }
-    var clone = exec(`git clone ${repo}`,{cwd:directory})
+    var clone
+    var cmd
+    if(uri){
+      cmd = `git clone --verbose --quiet ${uri}`
+    }else if(command){
+      cmd = command
+    }else{
+      callback("No uri or command given")
+      return false
+    }
+    var clone = exec(cmd,{cwd:directory})
     var output = []
     clone.stdout.on('data', function(a){
       logger.info(a)
@@ -127,8 +137,8 @@ Repo.create = function (repo, callback) {
 
     clone.on('exit',function(code){
       logger.debug('exit')
-      if(output.length>0 && code==0){
-        callback(null,output.join("\r\n"))
+      if(code==0){
+        callback(null,`repository created succesfully`)
       }else{
         callback(`\ncreating repository failed with code ${code}\n${output.join("\n")}`)
       }